extract function print_correct_errors_demo
must also extract color from terminal_color.h
