+++ memory-voting-prototype/makefile
@@ -26,11 +26,11 @@ test: all
 %.o: %.c
 	$(CC) $(CFLAGS) -c $< -o $@
 
-unit_tests: memory_correction.o testing.o
+unit_tests: memory_correction.o testing.o terminal_color.o
 	$(CC) $(CFLAGS) -o $@ unit_tests.c $^
 
 prototype: memory_correction.o testing.o
 	$(CC) $(CFLAGS) -o $@ prototype_voting.c $^
 
-demo: memory_correction.o testing.o util.o
+demo: memory_correction.o testing.o util.o terminal_color.o
 	$(CC) $(CFLAGS) -o $@ demo.c $^
