+++ memory-voting-prototype/terminal_color.c
@@ -1,0 +1,8 @@
+
+#include <unistd.h>
+
+#include "terminal_color.h"
+
+char* color(char* color) {
+  return isatty(STDOUT_FILENO) ? color : "";
+}
