+++ memory-voting-prototype/terminal_color.h
@@ -1,11 +1,7 @@
 #ifndef TERMINAL_COLOR_H
 #define TERMINAL_COLOR_H
 
-#include <unistd.h>
-
-char* color(char* color) {
-  return isatty(STDOUT_FILENO) ? color : "";
-}
+char* color(char* color);
 
 #define ANSI_COLOR_RED     color("\x1b[31m")
 #define ANSI_COLOR_GREEN   color("\x1b[32m")
