+++ memory-voting-prototype/util.c
@@ -3,6 +3,7 @@
 
 #include "util.h"
 #include "memory_correction.h"
+#include "terminal_color.h"
 
 void print_bits(char c) {
   // printf("%d ", c);
@@ -22,4 +23,17 @@ void print_data_copies_bits(struct data_copies_collection collection) {
     }
     printf("\n");
   }
+}
+
+void print_correct_errors_demo(int num_copies, size_t len, struct data_copies_collection collection) {
+  printf("%s\n Before Correction:\n\n%s", ANSI_COLOR_BRIGHT_CYAN, ANSI_COLOR_RESET);
+  print_data_copies_bits(collection);
+
+  int fixes = correct_errors(collection.data_copies, num_copies, (int)len);
+
+  printf("%s\n After Correction:\n\n%s", ANSI_COLOR_BRIGHT_CYAN, ANSI_COLOR_RESET);
+  print_data_copies_bits(collection);
+
+  printf("\n");
+  printf("%d fixes\n", fixes);
 }
